Int parsing in aKLPE query

+++ outlier_detection.py
@@ -78,7 +78,7 @@ def test_aKLPE(query, data, data_gs=None, k=DEFAULT_K, b=DEFAULT_B, far=DEFAULT_
         gs = train_aKLPE(data, k, b)
     
     for i in range(0, b):
-        subset = np.random.choice(n, np.floor(n/2), replace=False)
+        subset = np.random.choice(n, int(np.floor(n/2)), replace=False)
         g += calculate_G(distances[subset], k=k)
         
     g /= b
